vc4_bufmgr: fix time_t printf
Fixes:

  error: format ‘%ld’ expects argument of type ‘long int’, but argument 3 has type ‘time_t’ {aka ‘long long int’}

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/4279>

+++ src/gallium/drivers/vc4/vc4_bufmgr.c
@@ -100,7 +100,7 @@ vc4_bo_dump_stats(struct vc4_screen *screen)
                 struct timespec time;
                 clock_gettime(CLOCK_MONOTONIC, &time);
                 fprintf(stderr, "  now:               %ld\n",
-                        time.tv_sec);
+                        (long)time.tv_sec);
         }
 }
